Cleaned groupDetail and UserDetail Added everything into its own component and hooks. Almost. Didn't add groupDetail fetches, because it borked up the whole thing. Will fix later. Maybe.
Also fixed the GroupController patch and added an
description edit button to groupDetail.
