+++ frontend/src/components/Groups.js
@@ -1,10 +1,31 @@
-import React from 'react'
+import { Link } from 'react-router-dom';
 
 // On: 
-// - front page (Home or Index) when logged in and logged out
+// - UserDetail
 
-export default function Groups() {
+const Groups = ({ user }) => {
+  const groups = user.groups;
+
+  if (!groups) {
+    return <div>Loading...</div>;
+  }
+
   return (
-    <p>Groups listed here</p>
-  )
-}
+    <div className="user-tabs-content">
+      <h2>Ryhmät</h2>
+      <ul>
+        {groups.map(group => (  
+        <div key={group.id}>
+          <p>
+            <Link to={`/group/${group.id}`}>{group.groupName}</Link>
+          </p>
+          <p>{group.groupDescription}</p>
+          <hr></hr>
+        </div>
+      ))}
+      </ul>
+    </div>
+  );
+};
+
+export default Groups;
