+++ frontend/src/pages/GroupDetail.js
@@ -2,6 +2,9 @@ import React, { useEffect, useState } from 'react';
 import './GroupDetail.css';
 import { Link, useParams } from 'react-router-dom';
 import defaultGroupPicture from '../media/defaultGroupPicture.png';
+import GroupMembers from '../components/GroupMembers';
+import GroupReviews from '../components/GroupReviews';
+import GroupDetailTop from '../components/GroupDetailTop';
 //import { useAuth } from './auth-context'; // Import the useAuth hook from the auth-context file. change the file to the actual parameter
 
 function GroupDetail() {
@@ -89,34 +92,7 @@ const handleJoinLeaveGroup = () => {
   //Main content of the page
   return group ? (
     <div className="groupdetail-container">
-    <div className="groupdetail-top">
-      <div className="groupdetail-picture">
-        <img src={group.groupPicture} alt="Profile" />
-        <Link><p className='groupdetail-join-button'>Liity Ryhmään</p></Link>
-      </div>
-        <div className="groupdetail-bio">
-          <h2><Link to={`/group/${group.id}`}>{group.groupName}</Link></h2>
-          <p>{group.groupDescription}</p>
-          {
-          // Get the admin of the group and add a [ADMIN] tag after their name
-          }
-          {group.members.map(member => { 
-            if (member.isAdmin) {
-              return (
-                <React.Fragment key={member.userId.id}>
-                  <p>
-                    Admin: <Link to={`/users/${member.userId.id}`}>{member.userId.username}</Link>
-                  </p>
-                  {/* Below finished link when useAuth is implemented. Remove comment and curly brackets to use
-                  <Link to="#" onClick={handleJoinLeaveGroup}>{isMember ? 'Poistu ryhmästä' : 'Liity ryhmään'}</Link>
-                  */} 
-                </React.Fragment>
-              );
-            }      
-            return null;
-          })}
-        </div>
-      </div>
+      <GroupDetailTop group={group} />
       {
       // Tabs and their content
       }
@@ -156,20 +132,7 @@ const handleJoinLeaveGroup = () => {
           //Fetch a list of group members and display them. Also display admin tag if user is admin
           }
           {selectedTab === 'Members' && (
-            <div className="groupdetail-tabs-content">
-              <h2>Jäsenet</h2>
-              {group.members.map(member => (
-                <div key={member.groupMembersId}>
-                  <p>
-                    <Link to={`/users/${member.userId.id}`}>
-                      {member.userId.username}
-                      {member.isAdmin && <span> [ADMIN]</span>}
-                    </Link>
-                  </p>
-                  <p>{member.userId.userDescription}</p>
-                </div>
-              ))}
-            </div>
+            <GroupMembers group={group} />
           )}
           {selectedTab === 'Shared' && (
             <div className="groupdetail-tabs-content">
@@ -205,19 +168,7 @@ const handleJoinLeaveGroup = () => {
           //Fetch a list of reviews made by group members and display them. Also links to the movie on /movie/:id
           }
           {selectedTab === 'Reviews' && (
-            <div className="groupdetail-tabs-content">
-              <h2>Arvostelut</h2>
-              {reviews.map(review => (
-                <div key={review.id}>
-                  <p>
-                    <Link to={`/movie/${review.movie.id}`}>
-                      {review.movie.title}
-                    </Link>
-                  </p>
-                  <p>{review.content}</p>
-                </div>
-              ))}
-            </div>
+            <GroupReviews reviews={reviews} users={group.members} />
           )}
           {selectedTab === 'Finnkino' && (
             <div className="groupdetail-tabs-content">
@@ -257,20 +208,7 @@ const handleJoinLeaveGroup = () => {
           //Fetch a list of group members and display them. Also display admin tag if user is admin
           }
           {selectedTabLeft === 'Members' && (
-            <div className="groupdetail-tabs-content">
-              <h2>Jäsenet</h2>
-              {group.members.map(member => (
-                <div key={member.groupMembersId}>
-                  <p>
-                    <Link to={`/users/${member.userId.id}`}>
-                      {member.userId.username}
-                      {member.isAdmin && <span> [ADMIN]</span>}
-                    </Link>
-                  </p>
-                  <p>{member.userId.userDescription}</p>
-                </div>
-              ))}
-            </div>
+            <GroupMembers group={group} />
           )}
           {selectedTabLeft === 'Shared' && (
             <div className="groupdetail-tabs-content">
@@ -312,19 +250,7 @@ const handleJoinLeaveGroup = () => {
           //Fetch a list of reviews made by group members and display them. Also links to the movie on /movie/:id
           }
           {selectedTabRight === 'Reviews' && (
-            <div className="groupdetail-tabs-content">
-              <h2>Arvostelut</h2>
-              {reviews.map(review => (
-                <div key={review.id}>
-                  <p>
-                    <Link to={`/movie/${review.movie.id}`}>
-                      {review.movie.title}
-                    </Link>
-                  </p>
-                  <p>{review.content}</p>
-                </div>
-              ))}
-            </div>
+            <GroupReviews reviews={reviews} users={group.members} />
           )}
           {selectedTabRight === 'Finnkino' && (
             <div className="groupdetail-tabs-content">
