+++ frontend/src/pages/UserDetail.css
@@ -1,8 +1,6 @@
 .userdetail-container {
-  position: fixed;
   margin: 0em;
   width: 100%; 
-  height: 100%;
   overflow-y: auto;
   padding: 40px;
   background-color: var(--secondary-color); 
@@ -15,43 +13,14 @@
   align-items: flex-start; 
 }
 
-.user-top {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-}
-
-.user-profilepicture {
-  order: 1;
-}
-
-.user-details {
-  order: 2; 
-  margin-left: 20px;
-}
-
-.user-details h1{
-  text-transform: uppercase;
-}
-
-.user-details p{
-  font-size: small;
-}
-
-.user-bio {
-  order: 3; 
-  width: 100%; 
-  margin-top: 20px; 
-}
-
-.user-bio p{
-  width: 100%; 
-  font-size: small; 
+.user-tabs-container {
+  position: relative;
 }
 
 .user-tabs-content {
-  order: 6; 
+  height: 100%;
 }
+
 .user-tabs-content p{
   margin-top: 5px; 
   margin-bottom: 10px;
@@ -79,13 +48,18 @@
   display: flex;
   justify-content: space-between;
   width: 100%;
-  margin-top: 20px;
+  margin-top: 1px;
+  position: relative;
+  z-index: 1;
+  background-color: var(--secondary-color);
+  color: var(--primary-text-color); 
 }
 
 .user-content-left, .user-content-right {
   text-align: left;
   width: 50%;
   margin: 10px;
+  height: 100vh;
 }
 
 .user-tabs-container {
@@ -93,13 +67,15 @@
   display: flex;
   justify-content: space-between;
   width: 100%;
-  margin-top: 20px;
+  margin-top: 1px;
+  background-color: var(--background-color);
 }
 
 .user-tabs-left, .user-tabs-right {
   display: flex;
   flex-direction: row;
   background-color: var(--secondary-color);
+  min-width: max-content;
 }
 
 .user-tabs-left p, .user-tabs-right p {
@@ -114,32 +90,12 @@
 }
 
 .note-heading {
-  font-size: 1em; /* Adjust as needed */
+  font-size: 1em;
   font-weight: bold;
 }
 
 .active {
   background-color: var(--primary-color); 
-}
-.user-update-description {
-  order: 4;
-  justify-content: space-between;
-  width: 100%;
-  margin-top: 20px;
-  margin-bottom: 20px;
-}
-
-.user-update-description button{
-  float: right;
-  background-color: var(--primary-color);
-  border: 1px solid var(--primary-text-color);
-  border-radius: 5px;
-  padding: 5px;
-  margin-top: 5px;
-  text-transform: uppercase;
-  font-size:x-small;
-  width: 100px;
-  color: var(--primary-text-color);
 }
 
 @media (max-width: 600px) {
